Verify primary inductance value using MAXWELL
The simulated ETD44 with 1mm airgap with 59 turns gave an inductance value of 960uH. This is higher than what is designed for (900uH) but is alright since this inductance value would still ensure CCM operation. I also ran a 2mm air gap with ETD44 and saw the the inductabnce/turn^2 "AL" is expect. Also re organised the folder to separature PLECS model and MAXWELL models.