Created MAXWELL model for the ETD44 for verifying coreloss and Bsat
Was able to verify the Bsat of the ETD44 with 1.5mm airgap. Currently, the B is about 300mT at 4A (3.5A at fully power). Also made a transient model on maxwell to evaluate the coreloss and eventually, determine the temperature rise on the core. Also adding a CSV file that contains the fourier breakdown of the current waveforms which I still need to process in abe to become an input for the MAXWELL transient simulation.