Maxwell + Icepak simulation result for ETD44 (max temp rise is 25degC)
for ETD 44 1.5mm airgap, @ Vin = 293V , Fs = 100kHz, RL = 3.84 , D = 0.18. The core loss is about 0.2W , the copper loss is about 1W. The maximum temp rise is 25degC and it occurs around the copper widing. The highest temperature rise on the core is about 20degC around the center post. The rest of the core have around 10degC temp rise. The simulation ambient temperature is 10degC.