Copy V1.0 Schmatics & PCB before re-doing the schematics
I am redoing the schematics because the current schematic
looks is too messy because all the power and control section
are integrated into one sheet.

Another reason I am redoing schematic is because the current
schematics uses A1 or something very large which is hard to print.
I want to use A3 which is more readily available.